* Knghtktty whispers sweet nothings to Thyla (stuff about compilers and             graphics and ram upgrades and big hard drives...) <Thyla> oooooooooOOOOOOOOOO <Infinitas> Knghtktty: that's positively pornographic... * Thyla goes off into fits of ecstasy...

+++ dataprovider/src/main/java/qwde/dataprovider/models/IStockTicker.java
@@ -1,15 +1,14 @@
 package qwde.dataprovider.models;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public interface IStockTicker extends Comparable<IStockTicker>, Serializable {
-    BigDecimal getPrice();
+    double getPrice();
 
-    BigDecimal getHigh();
+    double getHigh();
 
-    BigDecimal getLow();
+    double getLow();
 
     Long getVolume();
