+++ dataprovider/src/main/java/qwde/dataprovider/models/StockTicker.java
@@ -2,18 +2,17 @@ package qwde.dataprovider.models;
 
 import qwde.dataprovider.util.DateUtil;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class StockTicker implements IStockTicker {
     public final String symbol;
-    public final BigDecimal closePrice;
-    public final BigDecimal highPrice;
-    public final BigDecimal lowPrice;
+    public final double closePrice;
+    public final double highPrice;
+    public final double lowPrice;
     public final Long volume;
     public final LocalDateTime timestamp;
 
-    public StockTicker(String symbol, BigDecimal price, BigDecimal highPrice, BigDecimal lowPrice, Long volume, LocalDateTime timestamp) {
+    public StockTicker(String symbol, double price, double highPrice, double lowPrice, Long volume, LocalDateTime timestamp) {
         this.symbol = symbol;
         this.closePrice = price;
         this.highPrice = highPrice;
@@ -23,17 +22,17 @@ public class StockTicker implements IStockTicker {
     }
 
     @Override
-    public BigDecimal getPrice() {
+    public double getPrice() {
         return this.closePrice;
     }
 
     @Override
-    public BigDecimal getHigh() {
+    public double getHigh() {
         return this.highPrice;
     }
 
     @Override
-    public BigDecimal getLow() {
+    public double getLow() {
         return this.lowPrice;
     }
